nvbios: fix sensor config attributes

+++ nvbios/nvbios.c
@@ -1562,7 +1562,7 @@ int main(int argc, char **argv) {
 			uint8_t hysteresis = type & 0xf;
 			const char *section_s = NULL, *threshold = NULL;
 			const char *correction_target = NULL;
-			int8_t correction_value = 0;
+			int16_t correction_value = 0;
 			uint16_t byte_low, byte_high;
 
 			if (id == 0x0)
